Refactor some tests and migrate globalrole tests
